update marbles lib to hackfest shim
