+++ chaincode/lib_marbles.go
@@ -21,10 +21,8 @@ package main
 
 import (
 	"encoding/json"
-	"errors"
 	"fmt"
-	"strconv"
-	"strings"
+	"errors"
 
 	"github.com/hyperledger/fabric/core/chaincode/shim"
 )
@@ -74,144 +72,5 @@ func get_complete_marble_index(stub shim.ChaincodeStubInterface) ([]string, erro
 		fmt.Println("marble index so far - ", completedMarbleIndex)
 	}
 
-	//change to array of bytes
-	//arrayAsBytes, _ := json.Marshal(completedMarbleIndex)
-	//return arrayAsBytes, nil
 	return completedMarbleIndex, nil
-}
-
-// ============================================================================================================================
-// delete_marble - remove a marble from state and from marble index
-// ============================================================================================================================
-func delete_marble(stub shim.ChaincodeStubInterface, args []string) ([]byte, error) {
-	fmt.Println("starting delete_marble")
-
-	if len(args) != 2 {
-		return nil, errors.New("Incorrect number of arguments. Expecting 2")
-	}
-
-	name := args[0]
-	authed_by_company := args[1]
-
-	//get the marble
-	marble, err := get_marble(stub, name)
-	if err != nil{
-		fmt.Println("Failed to find marble by name " + name)
-		return nil, err
-	}
-
-	//check authorizing company
-	if marble.Owner.Company != authed_by_company{
-		return nil, errors.New("The company '" + authed_by_company + "' cannot authorize deletion for '" + marble.Owner.Company + "'.")
-	}
-
-	//remove the marble
-	err = stub.DelState(name) //remove the key from chaincode state
-	if err != nil {
-		return nil, errors.New("Failed to delete state")
-	}
-
-	//get the marble index from owner
-	owner, err := get_owner(stub, marble.Owner.Username, marble.Owner.Company)
-	if err != nil {
-		fmt.Println("Failed to find owner - " + marble.Owner.Username + " " + marble.Owner.Company)
-		return nil, err
-	}
-
-	//remove marble from index
-	for i, val := range owner.Marbles {
-		fmt.Println(strconv.Itoa(i) + " - looking at " + val + " for " + name)
-		if val == name {                                                      //find the correct marble
-			fmt.Println("found marble")
-			owner.Marbles = append(owner.Marbles[:i], owner.Marbles[i+1:]...) //remove it
-			for x := range owner.Marbles {                                    //debug prints...
-				fmt.Println(string(x) + " - " + owner.Marbles[x])
-			}
-			break
-		}
-	}
-	var fullOwner = build_full_owner(marble.Owner.Username, marble.Owner.Company);
-	jsonAsBytes, _ := json.Marshal(owner)                                     //save new index
-	err = stub.PutState(fullOwner, jsonAsBytes)
-
-	fmt.Println("- end delete_marble")
-	return nil, nil
-}
-
-
-// ============================================================================================================================
-// Init Marble - create a new marble, store into chaincode state
-// ============================================================================================================================
-func init_marble(stub shim.ChaincodeStubInterface, args []string) ([]byte, error) {
-	var err error
-	fmt.Println("starting init_marble")
-
-	//   0       1       2     3      4                  5
-	// "asdf", "blue", "35", "bob", "united marbles", "united marbles"
-	if len(args) != 6 {
-		return nil, errors.New("Incorrect number of arguments. Expecting 6")
-	}
-
-	//input sanitation
-	if len(args[0]) <= 0 {
-		return nil, errors.New("1st argument must be a non-empty string")
-	}
-	if len(args[1]) <= 0 {
-		return nil, errors.New("2nd argument must be a non-empty string")
-	}
-	if len(args[2]) <= 0 {
-		return nil, errors.New("3rd argument must be a non-empty string")
-	}
-	if len(args[3]) <= 0 {
-		return nil, errors.New("4th argument must be a non-empty string")
-	}
-	if len(args[4]) <= 0 {
-		return nil, errors.New("5th argument must be a non-empty string")
-	}
-	name := args[0]
-	color := strings.ToLower(args[1])
-	username := strings.ToLower(args[3])
-	company := args[4]
-	authed_by_company := args[5]
-	size, err := strconv.Atoi(args[2])
-	if err != nil {
-		return nil, errors.New("3rd argument must be a numeric string")
-	}
-
-	//check authorizing company
-	if company != authed_by_company{
-		return nil, errors.New("The company '" + authed_by_company + "' cannot authorize creation for '" + company + "'.")
-	}
-
-	//check if marble already exists
-	marble, err := get_marble(stub, name)
-	if err == nil {
-		fmt.Println("This marble already exists - " + name)
-		fmt.Println(marble)
-		return nil, errors.New("This marble already exists - " + name) //all stop a marble by this name exists
-	}
-
-	//build the marble json string manually
-	str := `{"docType":"marble",  "name": "` + name + `", "color": "` + color + `", "size": ` + strconv.Itoa(size) + `, "owner": {"username": "` + username + `", "company": "` + company + `"}}`
-	err = stub.PutState(name, []byte(str))                     //store marble with id as key
-	if err != nil {
-		return nil, err
-	}
-
-	//get the marble index from owner
-	owner, err := get_owner(stub, username, company)
-	if err != nil {
-		fmt.Println("Failed to find owner - " + username + " " + company)
-		return nil, err
-	}
-
-	//append
-	var fullOwner = build_full_owner(username, company);
-	owner.Marbles = append(owner.Marbles, name)                //add marble name to index list
-	fmt.Println("! marble index - ", owner.Marbles)
-	jsonAsBytes, _ := json.Marshal(owner)
-	err = stub.PutState(fullOwner, jsonAsBytes)                //store name of marble
-
-	fmt.Println("- end init_marble")
-	return nil, nil
 }
